*** MAJOR CHANGE ***
-- catfacts
catfacts now is a single executed command to send a cat fact to a
certain number and message.

-- catfacts-client
This is the command line interactive or single executed program that
interfaces with catfacts-server. You will be able to add/edit/remove
the address book, and alarms of what to send to whom.

-- catfacts-server
This is a daemon that just manages an adress book and a queue of alarms
of catfacts to send to to people. It will try to focus on managing its
queue of alarms possibly in a database and its address book in a
database.

Too tired. Should have gone to bed earelier.